# Optimize recursive struct type comparison in `CompareAST`