--- a/src/incremental/compareAST.ml
+++ b/src/incremental/compareAST.ml
@@ -84,14 +84,7 @@ and mem_typ_acc (a: typ) (b: typ) acc = List.exists (fun p -> match p with (x, y
 and eq_typ_acc (a: typ) (b: typ) (acc: (typ * typ) list) =
   if Messages.tracing then Messages.tracei "compare" "eq_typ_acc %a vs %a\n" d_type a d_type b;
   (* if Messages.tracing then Messages.tracei "compare" "eq_typ_acc %a vs %a (%d)\n" d_type a d_type b (List.length acc); (* TODO: remove because always calls List.length *) *)
-  let r =
-  if mem_typ_acc a b acc || mem_typ_acc a b !global_typ_acc then (
-    if Messages.tracing then Messages.trace "compare" "in acc\n";
-    true
-  )
-  else (let acc = List.cons (a,b) acc in
-        let r =
-        match a, b with
+  let r = (match a, b with
         | TPtr (typ1, attr1), TPtr (typ2, attr2) -> eq_typ_acc typ1 typ2 acc && eq_list eq_attribute attr1 attr2
         | TArray (typ1, (Some lenExp1), attr1), TArray (typ2, (Some lenExp2), attr2) -> eq_typ_acc typ1 typ2 acc && eq_exp lenExp1 lenExp2 &&  eq_list eq_attribute attr1 attr2
         | TArray (typ1, None, attr1), TArray (typ2, None, attr2) -> eq_typ_acc typ1 typ2 acc && eq_list eq_attribute attr1 attr2
@@ -105,18 +98,26 @@ and eq_typ_acc (a: typ) (b: typ) (acc: (typ * typ) list) =
         | TNamed (tinf, attr), b -> eq_typ_acc tinf.ttype b acc (* Ignore tname, treferenced. TODO: dismiss attributes, or not? *)
         | a, TNamed (tinf, attr) -> eq_typ_acc a tinf.ttype acc (* Ignore tname, treferenced . TODO: dismiss attributes, or not? *)
         (* The following two lines are a hack to ensure that anonymous types get the same name and thus, the same typsig *)
-        | TComp (compinfo1, attr1), TComp (compinfo2, attr2) -> let res = eq_compinfo compinfo1 compinfo2 acc &&  eq_list eq_attribute attr1 attr2 in (if res && compinfo1.cname <> compinfo2.cname then compinfo2.cname <- compinfo1.cname); res
+        | TComp (compinfo1, attr1), TComp (compinfo2, attr2) ->
+          if mem_typ_acc a b acc || mem_typ_acc a b !global_typ_acc then (
+            if Messages.tracing then Messages.trace "compare" "in acc\n";
+            true
+          )
+          else (
+            let acc = (a, b) :: acc in
+            let res = eq_compinfo compinfo1 compinfo2 acc && eq_list eq_attribute attr1 attr2 in
+            if res && compinfo1.cname <> compinfo2.cname then
+              compinfo2.cname <- compinfo1.cname;
+            if res then
+              global_typ_acc := (a, b) :: !global_typ_acc;
+            res
+          )
         | TEnum (enuminfo1, attr1), TEnum (enuminfo2, attr2) -> let res = eq_enuminfo enuminfo1 enuminfo2 && eq_list eq_attribute attr1 attr2 in (if res && enuminfo1.ename <> enuminfo2.ename then enuminfo2.ename <- enuminfo1.ename); res
         | TBuiltin_va_list attr1, TBuiltin_va_list attr2 -> eq_list eq_attribute attr1 attr2
         | TVoid attr1, TVoid attr2 -> eq_list eq_attribute attr1 attr2
         | TInt (ik1, attr1), TInt (ik2, attr2) -> ik1 = ik2 && eq_list eq_attribute attr1 attr2
         | TFloat (fk1, attr1), TFloat (fk2, attr2) -> fk1 = fk2 && eq_list eq_attribute attr1 attr2
-        | _, _ -> false
-        in
-        if r then (* TODO: only do for nontrivial types? *)
-          global_typ_acc := (a, b) :: !global_typ_acc;
-        r
-        )
+        | _, _ -> false)
   in
   if Messages.tracing then Messages.traceu "compare" "eq_typ_acc %a vs %a\n" d_type a d_type b;
   r