To be merged after #358.

Closes #359.

### Fixes
First, there was a catch-all exception handler which was also catching `Stack_overflow` exceptions coming from recursive type comparison and hiding the problem. Those stack overflows just took time to happen.

Second, function argument type comparison was ignoring the `acc` structure that's supposed to avoid infinite recursion. A mix of recursive struct types with function pointer types to those structs caused the infinite recursion.

Thirdly, a global `acc` is added to reuse the recursive type equivalences between multiple globals, instead of having to go through all the trouble again.

Fourthly, the `acc` structure is only used for structs now, because that's what causes recursive types. By doing this the `acc` list is made significantly smaller, so checking it with `List.exists` is also much faster. There's no point in putting trivial equivalences like `(int, int)` into the structure because they're trivially computed from scratch.


### Performance improvement
Before (in #359):
```
Timings:
TOTAL                          104.756 s
  parse                           0.079 s
  convert to CIL                  0.093 s
  compareCilFiles                104.194 s
  analysis                        0.390 s
    global_inits                    0.078 s
    solving                         0.044 s
    verify                          0.021 s
```

After:
```
Timings:
TOTAL                           0.691 s
  parse                           0.074 s
  convert to CIL                  0.086 s
  compareCilFiles                 0.073 s
  analysis                        0.458 s
    global_inits                    0.074 s
    solving                         0.066 s
      S.Dom.equal                     0.001 s
    verify                          0.021 s
```

This makes `compareCilFiles` on the regression test **1400 times faster**!

And as you see, all these fixes were necessary to bring the comparison time of two unchanged ASTs down to be more like its parsing and conversion time, which is how it should be. Previously comparison of ASTs could be much bigger bottleneck than parsing itself!