@sim642 2021-09-28
If possible, the existing incremental benchmarks might be worth rerunning with this improvement. Both to see if this bottleneck could also appear there before or to notice possible issues introduced by this supposed fix.

Although it might be a bit difficult to redo that benchmark since this PR is on top of #358, which removes the Git-based incremental usage for the figlet benchmark.