@jerhard 2021-09-29
I let this run on the figlet repo and compared it to `master`. The run time of Goblint for the commits where things changed is pretty much the same (`master`: 118 seconds, this branch: `115` seconds.).