@sim642 2021-09-29
> I let this run on the figlet repo and compared it to `master`. The run time of Goblint for the commits where things changed is pretty much the same (`master`: 118 seconds (in total), this branch: 115 seconds; the time resolution was only on the levels of secons, however), so for the `figlet` repo it does not make so much of a difference.

Thanks for benchmarking it a bit! I think this is expected since figlet probably doesn't have these crazy recursive struct and function pointer types, which caused the biggest issue on Linux headers.